Rename 'newTrip' to 'signupTrip'. Rename folder 'my-profile' to 'myProfile'
