+++ package.json
@@ -74,9 +74,10 @@
     "react-datepicker": "0.27.0",
     "react-dom": "15.1.0",
     "react-redux": "4.4.5",
+    "react-redux-notifications": "0.2.1",
     "react-router": "2.5.0",
     "redux": "3.5.2",
-    "redux-form": "^5.2.5",
+    "redux-form": "5.2.5",
     "semantic-ui-css": "2.1.8"
   }
 }
