+++ src/commons/Sidebar.jsx
@@ -7,7 +7,7 @@ function Sidebar() {
             <div className="vertical-menu">
                 <SidebarMenuItem uri="/profile" itemName="My profile" />
                 <a href="#" className="item">My trips</a>
-                <SidebarMenuItem uri="/trips/new" itemName="Sign up for a trip" />
+                <SidebarMenuItem uri="/trips/signup" itemName="Sign up for a trip" />
                 <a href="#" className="item">Donations</a>
                 <a href="#" className="item">Messages</a>
                 <SidebarMenuItem uri="/admin/destinations" itemName="Admin Destinations" />
