+++ src/sections/signupTrip/SignupTripFormContainer.jsx
@@ -1,22 +1,22 @@
 import React from 'react';
 import { connect } from 'react-redux';
 
-import NewTripForm from './NewTripForm';
+import SignupTripForm from './SignupTripForm';
 import * as tripsApi from '../../api/trips';
 import * as destinationsApi from '../../api/destinations';
 
 
-class NewTripFormContainer extends React.Component {
+class SignupTripFormContainer extends React.Component {
     componentDidMount() {
         destinationsApi.getDestinations();
     }
     handleSubmit(data) {
-        tripsApi.postNewTrip(data);
+        tripsApi.postTrip(data);
     }
 
     render() {
         return (
-            <NewTripForm
+            <SignupTripForm
                 destinations={this.props.destinations}
                 onSubmit={e => { this.handleSubmit(e); }}
             />
@@ -28,8 +28,8 @@ const mapStateToProps = store => ({
     destinations: store.destinationState.destinations
 });
 
-NewTripFormContainer.propTypes = {
+SignupTripFormContainer.propTypes = {
     destinations: React.PropTypes.array.isRequired
 };
 
-export default connect(mapStateToProps)(NewTripFormContainer);
+export default connect(mapStateToProps)(SignupTripFormContainer);
